config/tmux: increase escape time to 1ms
The original intent of setting the escape time to 0 is to avoid weird
timing issues/bugs when switching between Insert/Normal mode on vim (and
other uses of escape keys).

When starting tmux from Windows Terminal (remote), however, there's a
bunch of garbage characters that are sent. Increasing the time to 1ms
alleviates this issue.

To be determined if it causes issues in vim, though.